requirements/tests.txt
pytest-asyncio
pytest-cov
pytest-env

# requesting test server
httpx

# for mocking aiohttp
aioresponses
